LayoutTests/fast/dom/shadow/shadow-root-js-api-expected.txt

Tests for ShadowRoot JS APIs. Can only run within DRT

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new WebKitShadowRoot() threw exception TypeError: Not enough arguments.
PASS new WebKitShadowRoot(undefined) threw exception Error: HIERARCHY_REQUEST_ERR: DOM Exception 3.
PASS new WebKitShadowRoot(null) threw exception Error: HIERARCHY_REQUEST_ERR: DOM Exception 3.
PASS new WebKitShadowRoot(1) threw exception Error: HIERARCHY_REQUEST_ERR: DOM Exception 3.
PASS shadowHost is shadowRoot.host
PASS new WebKitShadowRoot(shadowHost) threw exception Error: HIERARCHY_REQUEST_ERR: DOM Exception 3.
The previous test should not have side effects.
PASS shadowHost is shadowRoot.host
PASS successfullyParsed is true

TEST COMPLETE
